Rename stackdriver->otlp
